fix try-except in base_dao

+++ src/daos/base_dao.py
@@ -1,6 +1,7 @@
 from typing import Generic, TypeVar
 
-from pydantic import BaseModel
+from fastapi import HTTPException
+from pydantic import BaseModel, ValidationError
 from sqlalchemy import select
 from sqlalchemy.exc import SQLAlchemyError, NoResultFound
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -18,15 +19,19 @@ class BaseDAO(Generic[T]):
 
     @classmethod
     async def add(cls, session: AsyncSession, model: BaseModel) -> int:
-        request = cls.model(**model.model_dump())
+        try:
+            model_data = model.model_dump(exclude_unset=True)
+            request = cls.model(**model_data)
+        except ValidationError as e:
+            raise ValueError('Invalid data') from e
+
         try:
             session.add(request)
             await session.commit()
             await session.refresh(request)
         except SQLAlchemyError as e:
             await session.rollback()
-            err = str(e)
-            raise err
+            raise e
 
         return request.id
 
@@ -35,7 +40,13 @@ class BaseDAO(Generic[T]):
         try:
             query = select(cls.model)
             result = await session.execute(query)
+
+            if result is None:
+                raise HTTPException(status_code=404, detail="Not found")
+
             return result.scalars().all()
+        except HTTPException as e:
+            raise
         except SQLAlchemyError as e:
             await session.rollback()
             err = str(e)
@@ -45,7 +56,13 @@ class BaseDAO(Generic[T]):
     async def get_by_id(cls, session: AsyncSession, item_id: int) -> dict | None:
         try:
             query = await session.get(cls.model, item_id)
+
+            if query is None:
+                raise HTTPException(status_code=404, detail="Item not found")
+
             return query.info
+        except HTTPException as e:
+            raise
         except SQLAlchemyError as e:
             err = str(e)
             raise err
@@ -53,28 +70,32 @@ class BaseDAO(Generic[T]):
     @classmethod
     async def delete(cls, session: AsyncSession, item_id: int) -> bool:
         try:
-            query = session.get(cls.model, item_id)
+            query = await session.get(cls.model, item_id)
+
+            if query is None:
+                raise HTTPException(status_code=404, detail="Item not found")
+
             await session.delete(query)
             await session.commit()
-            await session.refresh(query)
             return True
-        except SQLAlchemyError as e:
+        except HTTPException as e:
+            raise
+        except Exception as e:
             await session.rollback()
-            err = str(e)
-            raise err
+            raise SQLAlchemyError(f'Error deleting: {str(e)}') from e
 
     @classmethod
     async def patch(cls, session: AsyncSession, item_id: int, item_data: BaseModel) -> int | None:
-        item_dict = item_data.model_dump()
+        item_dict = item_data.model_dump(exclude_unset=True)
         try:
-            req = session.get(cls.model, item_id)
+            req = await session.get(cls.model, item_id)
 
             for key, value in item_dict.items():
                 setattr(req, key, value)
 
             await session.commit()
             await session.refresh(req)
-            return item_dict.id
+            return req.info
         except NoResultFound:
             return None
         except SQLAlchemyError as e:
